+++ apps/webapp/app/services/events/deliverEvent.server.ts
@@ -1,6 +1,6 @@
 import type { EventDispatcher, EventRecord } from "@trigger.dev/database";
 import type { EventFilter } from "@trigger.dev/core";
-import { EventFilterSchema } from "@trigger.dev/core";
+import { EventFilterSchema, eventFilterMatches } from "@trigger.dev/core";
 import { $transaction, PrismaClientOrTransaction, prisma } from "~/db.server";
 import { logger } from "~/services/logger.server";
 import { workerQueue } from "../worker.server";
@@ -124,29 +124,6 @@ export class EventMatcher {
   }
 
   public matches(filter: EventFilter) {
-    return patternMatches(this.event, filter);
+    return eventFilterMatches(this.event, filter);
   }
-}
-
-function patternMatches(payload: any, pattern: any): boolean {
-  for (const [patternKey, patternValue] of Object.entries(pattern)) {
-    const payloadValue = payload[patternKey];
-
-    if (Array.isArray(patternValue)) {
-      if (patternValue.length > 0 && !patternValue.includes(payloadValue)) {
-        return false;
-      }
-    } else if (typeof patternValue === "object") {
-      if (Array.isArray(payloadValue)) {
-        if (!payloadValue.some((item) => patternMatches(item, patternValue))) {
-          return false;
-        }
-      } else {
-        if (!patternMatches(payloadValue, patternValue)) {
-          return false;
-        }
-      }
-    }
-  }
-  return true;
 }