+++ packages/cli/src/utils/detectNextJsProject.ts
@@ -1,15 +1,29 @@
 import fs from "fs/promises";
 import pathModule from "path";
+import { readPackageJson } from "./readPackageJson.js";
 
 /** Detects if the project is a Next.js project at path  */
 export async function detectNextJsProject(path: string): Promise<boolean> {
-  // Checks for the presence of a next.config.js file
-  try {
-    // Check if next.config.js file exists in the given path
-    await fs.access(pathModule.join(path, "next.config.js"));
+  const hasNextConfigFile = await detectNextConfigFile(path);
+  if (hasNextConfigFile) {
     return true;
-  } catch (error) {
-    // If next.config.js file doesn't exist, it's not a Next.js project
+  }
+
+  return await detectNextDependency(path);
+}
+
+async function detectNextConfigFile(path: string): Promise<boolean> {
+  return fs
+    .access(pathModule.join(path, "next.config.js"))
+    .then(() => true)
+    .catch(() => false);
+}
+
+async function detectNextDependency(path: string): Promise<boolean> {
+  const packageJsonContent = await readPackageJson(path);
+  if (!packageJsonContent) {
     return false;
   }
+
+  return packageJsonContent.dependencies?.next !== undefined;
 }