+++ integrations/resend/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @trigger.dev/resend
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [84126808]
+  - @trigger.dev/integration-kit@2.0.7
+  - @trigger.dev/sdk@2.0.7
+
+## 2.0.6
+
+### Patch Changes
+
+- @trigger.dev/integration-kit@2.0.6
+- @trigger.dev/sdk@2.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- @trigger.dev/integration-kit@2.0.5
+- @trigger.dev/sdk@2.0.5
+
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies [96384991]
+  - @trigger.dev/sdk@2.0.4
+  - @trigger.dev/integration-kit@2.0.4
+
 ## 2.0.3
 
 ### Patch Changes