+++ packages/trigger-sdk/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @trigger.dev/sdk
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [fa3a22eb]
+  - @trigger.dev/core@2.0.7
+
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [59075f5f]
+  - @trigger.dev/core@2.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- @trigger.dev/core@2.0.5
+
+## 2.0.4
+
+### Patch Changes
+
+- 96384991: Adding the validate endpoint action to be able to add an endpoint first in the dashboard
+- Updated dependencies [96384991]
+  - @trigger.dev/core@2.0.4
+
 ## 2.0.3
 
 ### Patch Changes